==> pipeline/07_POST_LINKAGE_QA_NoteBook.ipynb <==
# Clerical Search to estimate false negative rate

# Packages
import pandas as pd
import numpy as np
import functools
import re

# Read in the census & PES data
CEN = pd.read_csv(DATA_PATH + 'census_cleaned.csv', index_col=False)
print("Census read in")
PES = pd.read_csv(DATA_PATH + 'pes_cleaned.csv', index_col=False)  
print("PES read in")

# Sample of unmatched PES records (just a list of PES IDs)
unmatched_PES = pd.read_csv(DATA_PATH + 'Stage_6_Unmatched_PES_Sample.csv')

### 1) Open csv above (Stage_6_Unmatched_PES_Sample) in Excel and keep open whilst matching
### 2) Create new column 'MATCH' which will contain a 1 if a match is found in the census, otherwise 0
### 3) Create new column 'CENSUS_ID' which will contain the census puid if a match is found, otherwise 0
### 4) Take an unmatched PES record from the sample and view all variables of interest

# Variables of interest when presenting PES / CEN records
variables = ['puid', 'hhid', 'EAid', 'DSid', 'names', 'year_birth', 'birth_month', 'sex', 'relationship_hh']
pes_variables = [x + '_pes' for x in variables]
cen_variables = [x + '_cen' for x in variables]

# View unmatched PES record - this is the record we want to try and find in the census
target_record = PES[PES.puid_pes == '111'][pes_variables]
pd.melt(target_record)

### 5) View all other records from the PES household


# View all records from the household (matched or unmatched)
target_household_ID = target_record['hhid_pes'][0]
target_household = PES[PES.hhid_pes == target_household_ID][pes_variables]
target_household

### 6) Search for a match in the full census dataset using a selection of different filters

# Function to combine and apply multiple filters to the census dataset
def conjunction(*conditions):
    return functools.reduce(np.logical_and, conditions)

### Create a list of conditions to search the Census dataset with here

# Comment out filters not used and add in any extras if you find they work better at finding matches
c1 = CEN.first_name_cen == 'CHARLIE'
c2 = CEN.last_name_cen == 'T'
c3 = CEN.year_birth_cen.between(1990, 1995)
c4 = CEN.birth_month_cen.between(5, 12)
c5 = CEN.sex_cen == '1'
c6 = CEN.EAid_cen == "123"

# Apply chosen filters to census dataset
conditions_list = [c1, c2, c3, c4, c5, c6]
census_candidates = CEN[conjunction(*conditions_list)][cen_variables]
print("Search has produced {} potential census match/matches".format(len(census_candidates)))

# View candidates
census_candidates

### 7) If you think you may have found a census record that matches to the PES record, you may want to view the whole census household

# View full household for person of interest
census_household_ID = CEN[CEN.puid_cen == '111']['hhid_cen'][0]
census_household = CEN[CEN.hhid_cen == census_household_ID][cen_variables]
census_household

### 8) Other filters that you may want to try in step 6 - just add them to "conditions_list" ([c1,c2,c3.....cN])

# N-grams e.g. first 2 letters of first name / last 5 letters of last name
CEN.first_name_cen.str[0:2] == 'CH'
CEN.first_name_cen.str[-5:] == 'ARLIE'

# Missing value filter
CEN.last_name_cen.isnull()

# Filter multiple possible first names
CEN.first_name_cen.isin(['CHARLIE', 'CHARLES', 'CHAZ'])

# Wildcard - Search for a name and allow for one or more characters where the .+ is
# e.g. CHARLIE, CHABLIE, CHARLLE, CHRALIE, CHALRIE, CHAARLIE etc.
def wildcard(string):
    if re.search('CH.+IE',string): return True
    else: return False
  
# Apply wildcard filter
CEN.first_name_cen.apply(wildcard)

# Other variables you could filter on:    
    # Head of Household
    # Relationship to Head of Household 
    # Marital Status
    # Different levels of geography e.g. HH, EA, District

PES = pd.DataFrame({'puid_pes': ['111', '222', '333', '444', '555'],
                    'hhid_pes': ['11', '11', '11', '22', '22'],
                    'names_pes': ['CHARLIE T ', 'STEVE X', 'JOHN P', 'BOB Y', 'PETE'],
                    'first_name_pes': ['CHARLIE', 'STEVE', 'JOHN', 'BOB', 'PETE'],
                    'last_name_pes': ['T', 'X', 'P', 'Y', None],
                    'year_birth_pes': [1993, 1999, 1992, 2000, 1970],
                    'birth_month_pes': [7, 8, 1, 12, 6],
                    'sex_pes': ['1', '2', '2', '2', '1'],
                    'EAid_pes': ['123', '456', '789', '123', '456'],
                    'DSid_pes': ['1', '4', '7', '1', '4'],
                    'relationship_hh_pes': ['1', '2', '2', '1', '3']})

CEN = pd.DataFrame({'puid_cen': ['111', '222', '333', '444', '555'],
                    'hhid_cen': ['11', '11', '11', '22', '22'],
                    'names_cen': ['CHARLIE T ', 'STEVE X', 'JOHN P', 'BOB Y', 'PETE'],
                    'first_name_cen': ['CHARLIE', 'STEVE', 'JOHN', 'BOB', 'PETE'],
                    'last_name_cen': ['T', 'X', 'P', 'Y', None],
                    'year_birth_cen': [1993, 1999, 1992, 2000, 1970],
                    'birth_month_cen': [7, 8, 1, 12, 6],
                    'sex_cen': ['1', '2', '2', '2', '1'],
                    'EAid_cen': ['123', '456', '789', '123', '456'],
                    'DSid_cen': ['1', '4', '7', '1', '4'],
                    'relationship_hh_cen': ['1', '2', '2', '1', '3']})
